finished I guess?

+++ woo/app/main/DoShowGlobalBalance.java
@@ -4,7 +4,6 @@ import pt.tecnico.po.ui.Command;
 import pt.tecnico.po.ui.DialogException;
 import pt.tecnico.po.ui.Input;
 import woo.core.StoreManager;
-//FIXME import other classes
 
 /**
  * Show global balance.
@@ -18,8 +17,8 @@ public class DoShowGlobalBalance extends Command<StoreManager> {
   @Override
   public final void execute() {
     _display.popup(Message.currentBalance(
-    	Math.round(_receiver.getAvailableBalance()),
-    	Math.round(_receiver.getAccountingBalance())    	
+    	(int)Math.round(_receiver.getAvailableBalance()),
+    	(int)Math.round(_receiver.getAccountingBalance())    	
     ));
   }
 }
